Implement Super-Resolution Dataset
This custom PyTorch Dataset loads images from a
folder, performs several image operations upon
them, and then returns patches to the caller.
To avoid expensive disk reads/writes, all of the
images are cached in memory in compressed form,
and then decompressed and cached as the iterator
iterates over them.